Float should map to warehouse FLOAT type (#956)
Float is a valid spark Datatype. There's no
reason to map floats to the larger double type

+++ elasticgraph-warehouse/lib/elastic_graph/warehouse/schema_definition/api_extension.rb
@@ -40,7 +40,7 @@ module ElasticGraph
           "Cursor" => "STRING",
           "Date" => "DATE",
           "DateTime" => "TIMESTAMP",
-          "Float" => "DOUBLE",
+          "Float" => "FLOAT",
           "ID" => "STRING",
           "Int" => "INT",
           "JsonSafeLong" => "BIGINT",
